fix: divide classes AdminController, AdminService into several classes
