+++ main-service/src/main/java/ru/yandex/practicum/explore/request/model/Request.java
@@ -10,7 +10,6 @@ import java.time.LocalDateTime;
 @Setter
 @Getter
 @EqualsAndHashCode
-@ToString
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
